requirements.txt
selenium==4.17.0
pytest==8.2.2
pytest-html==4.2.1
openpyxl==3.1.3
PyYAML==6.0
pandas==2.1.1
